requirements.txt
streamlit
mtcnn
opencv-python
pillow
numpy
tensorflow==2.15.0
keras-facenet
scipy
